% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHCprops}
\alias{CHCprops}
\title{Chlorinated Hydrocarbon properties}
\format{Data frame with 4 named rows:
\describe{
  \item{solubility}{solubility in water close to 20 degrees C, in
   grams/l}
  \item{density}{free phase density close to 20 degrees C, in kg/m3}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pccompound}
}
\usage{
CHCprops
}
\description{
Properties of selected chlorinated solvents that are relevant to the
 DNAPL source term model, namely solubility and density.
}
\details{
There is significant variation in the figures for TeCM solubility.
 Values are for temperatures between 20 and 25 degrees celcius.
}
\keyword{datasets}

