% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_model.R
\name{cstG.DNmodel}
\alias{cstG.DNmodel}
\title{Constant power mass-flux relationship model}
\usage{
cstG.DNmodel(wg, wpm, hp, Gamma, Srn, Srw, phi, rho, Cs, hL,
  NLAY = length(hL))
}
\arguments{
\item{wg}{numeric \code{[1]} or \code{[NLAY]};
ganglia width}

\item{wpm}{numeric \code{[1]} or \code{[NLAY]};
maximum pool width (may be \code{Inf} in the lowest layer, in which case
 no DNAPL is allowed to escape)}

\item{hp}{numeric \code{[1]} or \code{[NLAY]};
pool height, or total height of pools in a layer}

\item{Gamma}{numeric \code{[1]} or \code{[NLAY]};
empirical source depletion parameter, positive; small values (<1) imply
 a more persistent source term}

\item{Srn, Srw}{numeric \code{[1]};
residual saturations of NAPL and water}

\item{phi}{numeric \code{[1]} or \code{[NLAY]};
bulk porosity of each layer}

\item{rho}{numeric \code{[1]};
solvent density (ensure consistent units, probably kg/m^3)}

\item{Cs}{numeric \code{[1]};
solvent solubility in water (ensure consistent units, probably kg/m^3
 which is the same as g/l)}

\item{hL}{numeric \code{[1]} or \code{[NLAY]};
the height of each layer; note that higher layers will contain more NAPL
 mass}

\item{NLAY}{integer \code{[1]};
number of layers in the DNAPL model}
}
\description{
Constant power mass-flux relationship model
}

