% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_model.R
\name{DDpg.DNmodel}
\alias{DDpg.DNmodel}
\title{Dual-domain Pool and Ganglia model}
\usage{
DDpg.DNmodel(wg, wpm, hp, Srn, Srw, phi, rho, Cs, hL, NLAY = length(hL),
  poolGamma = 1)
}
\arguments{
\item{wg}{numeric \code{[1]} or \code{[NLAY]};
ganglia width}

\item{wpm}{numeric \code{[1]} or \code{[NLAY]};
maximum pool width (may be \code{Inf} in the lowest layer, in which case
 no DNAPL is allowed to escape)}

\item{hp}{numeric \code{[1]} or \code{[NLAY]};
pool height, or total height of pools in a layer}

\item{Srn}{numeric \code{[1]};
residual saturations of NAPL and water}

\item{Srw}{numeric \code{[1]};
residual saturations of NAPL and water}

\item{phi}{numeric \code{[1]} or \code{[NLAY]};
bulk porosity of each layer}

\item{rho}{numeric \code{[1]};
solvent density (ensure consistent units, probably kg/m^3)}

\item{Cs}{numeric \code{[1]};
solvent solubility in water (ensure consistent units, probably kg/m^3
 which is the same as g/l)}

\item{hL}{numeric \code{[1]} or \code{[NLAY]};
the height of each layer; note that higher layers will contain more NAPL
 mass}

\item{NLAY}{integer \code{[1]};
number of layers in the DNAPL model}

\item{poolGamma}{numeric \code{[1]} or \code{[NLAY]};
the empirical depletion power to be applied to pools, analogous to
 \code{Gamma} in \code{\link{cstG.DNmodel}} which applies to the bulk
 NAPL}
}
\value{
a \link{DNAPLmodel} S4 object
}
\description{
This model has separate domains for ganglia and pools.  Ganglia are
 flushed efficiently (effectively \code{Gamma = Inf}) and pools are
 flushed inefficiently, with a \code{poolGamma} depletion parameter that
 defaults to 1.
}
