% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_model.R
\docType{class}
\name{DNAPLmodel-class}
\alias{DNAPLmodel}
\alias{DNAPLmodel-class}
\title{DNAPL distribution and dissolution model}
\description{
This is the structure of a DNAPL model used by \code{\link{DNST}} to
 specify the behaviour and dissolution of DNAPLs layer by layer.
}
\details{
Use the \code{\link{DNAPLmodel.check}} function for a thorough analysis
 of the DNAPL model, to make sure that it is set up correctly.
}
\section{Slots}{

\describe{
\item{\code{NLAY}}{integer;
number of layers in the DNAPL model}

\item{\code{hL}}{numeric;
height of each DNAPL model layer}

\item{\code{params}}{list;
a list of input parameters specific to the model}

\item{\code{domains}}{vector of character strings;
names of the mass domains, one of which will probably be "plume"; other
 common examples include "NAPL" (lumped NAPL), "pool", "ganglia",
 "vapour", "fractureNAPL", ...}

\item{\code{domain1}}{character string;
the name of the domain to which the infiltrating DNAPL is sent initially;
 must feature in \code{domains}}

\item{\code{mdmax}}{numeric matrix;
with rows representing layers and columns (named) representing domains,
 layer-by-layer values for the maximum mass in each domain; Inf is
 permitted and is often used for the plume domain}

\item{\code{mdredist}}{list;
elements should be of class \code{\link{Mredistribution}}; number of
 elements is arbitrary as each element represents a mass redistribtion
 process}

\item{\code{spill.to}}{data.frame;
one named row for each domain, except for domains with infinite capacity
 (as in \code{mdmax}) in every layer; this data frame redirects mass
 that exceeds the capacity for a domain in a given layer to a new domain
 or a new layer.  The data frame should have columns:\cr
 \code{$domain} chr/ factor: domain to which overspilt mass from this
  domain (in the \code{row.names}) should be directed; may be itself if
  \code{layer != 0}; for example excess mass in a "ganglia" domain may
  spill to the "pool" domain in the same layer, (in which case
  \code{layer} would be 0), or excess mass in a "NAPL" domain may spill
  to the "NAPL" domain in the underlying layer (in which case
  \code{layer} would be 1)
 \code{$layer} int: 1 for directing to the layer below; 0 for the same
  layer or (rarely) -1 for the layer above
 optional \code{$domain2} chr/ factor: a second choice domain}
}}

