% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_solver.R
\docType{class}
\name{DNAPLSourceTerm-class}
\alias{DNAPLSourceTerm}
\alias{DNAPLSourceTerm-class}
\title{DNAPL Source Term result}
\value{
DNAPLSourceTerm object; the results from \code{\link{DNST}}
}
\description{
Format for the results from \code{\link{DNST}}
}
\section{Slots}{

\describe{
\item{\code{Jspill}}{numeric \code{[NTS]};
DNAPL spill flux by each time step}

\item{\code{Jeffluent}}{numeric matrix \code{[NLAY, NTS]};
source term effluent flux to the plume, by layer (matrix row) and time
 step (matrix column); if there was no plume domain, this is an empty
 matrix}

\item{\code{M}}{numeric array \code{[NLAY, NTS, Ndom]};
source zone mass by layer (dim1), time step (dim2) and mass domain
 (dim3)}

\item{\code{Mbot,Mtop}}{numeric \code{[NTS]};
cumulative mass escaped from the bottom and top of the DNAPL model}

\item{\code{time}}{numeric \code{[NTS]};
time values at the end of each time step}

\item{\code{DNAPLmodel}}{DNAPLmodel;
The input DNAPL model (see \code{\link{DNAPLmodel}})}

\item{\code{imbalance}}{numeric \code{[NTS]};
mass imbalance by time step; mass should be completely accounted for, so
 any \code{imbalance} asides from machine imprecision constitutes an
 error}

\item{\code{description}}{character string;
user-supplied information specific to the model, for future reference}
}}

