% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_master.R
\name{DNST_MASTER}
\alias{DNST_MASTER}
\title{DNAPL source term master}
\usage{
DNST_MASTER(result.file, description, x, y = NULL, mfdata, contnt, DNmodel,
  Dm.pars, pu, start.t = td::td(1, 1, 1925), end.t, dt = 20, fw = 1,
  fi = 1, fsphist = data.frame(year = c(1974, 1990), f = c(1, 0)),
  ndlpmfl = 1L, bph = 0, z0 = "base")
}
\arguments{
\item{result.file}{character string;
file to save results to (should end with extension \code{".rds"})}

\item{description}{character string;
information about the model and what it represents; this is saved with
 the results for your record}

\item{x}{x and y co-ordinates of the spill with the same origin and units as
 \code{mfdata};
used to find the grid cell from which to read MODFLOW results;
 alternatively, if \code{mfdata} is not used, x and y can simply be
 given as information to be stored with the model}

\item{y}{x and y co-ordinates of the spill with the same origin and units as
 \code{mfdata};
used to find the grid cell from which to read MODFLOW results;
 alternatively, if \code{mfdata} is not used, x and y can simply be
 given as information to be stored with the model}

\item{mfdata}{NetCDF object;
MODFLOW data in NetCDF format (see \code{\link[Rflow]{GW.nc}}),
 from which the transient horizontal Darcy velocity through the source
 zone is determined}

\item{contnt}{\code{"TCE"}, \code{"PCE"}, \code{"TCA"} or {"TeCM"}}

\item{DNmodel}{function (as object or character string);
function to create object of class \link{DNAPLmodel}, such as
 \code{\link{cstG.DNmodel}}}

\item{Dm.pars}{named or correctly ordered list;
arguments for \code{DNmodel} (see \code{\link[base]{formals}}),
excluding hL, NLAY, Cs and rho; \code{\link[base]{do.call}} is used}

\item{pu}{numeric \code{[1]};
the peak usage rate at the site ever - the national usage rate is
 normalised to give it a maximum value of \code{pu}; because \code{cons}
 is normalised, \code{pu} needn't be in the same units - it should be in
 the desired units of output and kg/day is suggested}

\item{start.t}{numeric \code{[1]};
start and end time of the model, in days since 30/12/1899}

\item{end.t}{numeric \code{[1]};
start and end time of the model, in days since 30/12/1899}

\item{dt}{numeric \code{[1]};
time step length, in days, usually in the order of tens of days}

\item{fw}{numeric \code{[1]};
proportion of imported solvent that becomes waste (between 0 and 1, 1 is
 suggested)}

\item{fi}{numeric \code{[1]};
proportion of spilt solvent that infiltrates to water table (between 0
 1, 1 is suggested)}

\item{fsphist}{data.frame:\cr
\code{$year} (num)\cr
\code{$f} (num): fraction of imported solvent that is spilt, rather than
 disposed of safely\cr
the transient safe disposal fraction of solvent by the site}

\item{ndlpmfl}{integer \code{[1]} or \code{[mfNLAY]}, where \code{mfNLAY} is the number
 of layers in the MODFLOW model;
number of DNAPL model layers per MODFLOW layer (including any which are
 always dry at the location, although the values for these layers won't
 make any difference)}

\item{z0}{numeric \code{[1]} or \code{"base"};
elevation of the base of the DNAPL model, with the same datum as
 \code{mfdata}; \code{"base"} instructs to use the bottom elevation of
 the lowest layer at \code{x,y}, read from \code{mfdata}}
}
\value{
\link{DNAPLSourceTerm} object, which is also written to
 \code{result.file}.  The results are returned invisibly as they will be
 too complex to be usefully printed to screen.  Assign the results to an
 object or read from \code{result.file} using
 \code{\link[base]{readRDS}}.
}
\description{
A convenient master function for DNAPL source terms when a MODFLOW
 data set and the contaminant data frames in the package are being used.
}

