% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_model.R
\docType{class}
\name{Mredistribution-class}
\alias{Mredistribution}
\alias{Mredistribution-class}
\title{Mass Redistribution function}
\description{
Mass Redistribution function
}
\details{
The \code{flux} function must have all four arguments even if in fact
 they are not all used.  The function may also use global variables that
 will exist in the \code{\link{DNST}} solver function.  Typically,
 \code{qh} will feature somewhere in some functions, as the dissolution
 mass flux, for example, will depend on the water speed through the
 source zone.  Some functions may also use the \code{M} array in order
 to include some dependency on the source zone history.
}
\section{Slots}{

\describe{
\item{\code{from}}{character string;
the domain from which mass is taken}

\item{\code{to}}{character string;
the domaint to which mass is sent}

\item{\code{flux}}{function;
a function whose arguments are:\cr
\code{fromM}: mass of \code{from} domain\cr
\code{toM}: mass of \code{to} domain\cr
\code{LAY}: the layer of the DNAPL model\cr
\code{time}: in days since 30/12/1899\cr
and whose output represents the rate of mass transfer between the
 \code{from} and \code{to} domains (negative outputs are permitted and
 imply a flux the other way; in this way an equilibrium may be set up)}
}}

