% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overspill}
\alias{overspill}
\title{excess mass in domains}
\usage{
overspill(TS, mdmax)
}
\arguments{
\item{TS}{time step number}

\item{mdmax}{from DNAPLmodel}
}
\value{
matrix with colnames corresponding to DNAPLmodel@domains and nlay rows
}
\description{
excess mass in domains
}

