% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm.R
\name{C0_pool_XsectA}
\alias{C0_pool_Hunt88}
\alias{C0_pool_XsectA}
\title{Flow-normal cross-sectional area PDM}
\usage{
C0_pool_XsectA(Cs, hp, hL)

C0_pool_Hunt88(Cs, wp0, hL, aV = wp0/1000, Np = 1L)
}
\arguments{
\item{Cs}{numeric [];
solubility}

\item{hp}{numeric [];
pool height}

\item{hL}{numeric [];
layer thickness}

\item{wp0}{numeric [];
historic maximum of pool width}

\item{aV}{numeric [1];
vertical dispersivity, in units of length}

\item{Np}{integer [1];
number of pools within the layer; more pools will result in a greater
 average effluent concentration as no point will be so far from a pool}
}
\value{
numeric [];
C0, the historic maximum of effluent concentration averaged through the
 layer

numeric [];
C0, the historic maximum of effluent concentration averaged through the
 layer
}
\description{
integral of \code{Cs*erfc(z/(2*sqrt(ax*wp0)))} between 0 and \code{hp},
 or else lots of smaller pools if \code{Np} isn't 1
}
\note{
It is assumed that dispersion only occurs above the pool, as pools rest
 on top of aquitards, so hydrodynamic dispersion below the pool should
 be fairly small, instead just molecular diffusion should occur into the
 underlying aquitard.  If you want to dispense with this assumption,
 then double the value for \code{Np}, as the two-way dispersion from one
 pool at the bottom of the layer will result in the same average
 concentration as the one-way dispersion from two pools, one at the
 bottom and one in the middle.
}

