% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNST_model.R
\name{pool_mpua}
\alias{pool_mpua}
\title{Mass per unit area of pool}
\usage{
pool_mpua(hp, phi, rho, Srn, Srw, mode = c("uniform", "linear",
  "exponential")[1L])
}
\arguments{
\item{hp}{numeric \code{[1]} or \code{[NLAY]};
pool height, or total height of pools in a layer}

\item{phi}{numeric \code{[1]} or \code{[NLAY]};
bulk porosity of each layer}

\item{rho}{numeric \code{[1]};
solvent density (ensure consistent units, probably kg/m^3)}

\item{Srn}{numeric \code{[1]};
residual saturations of NAPL and water}

\item{Srw}{numeric \code{[1]};
residual saturations of NAPL and water}

\item{mode}{character [1];
nature of saturation variation with height above pool base, one of
 \code{"uniform"}, \code{"linear"} or \code{"exponential"}; for
 \code{"uniform"}, it is taken that the pool has saturation
 \code{1 - Srw} and for the other two it is taken that the pool has
 saturation \code{1 - Srw} at the base and \code{Srn} at the top}
}
\value{
numeric[]
}
\description{
Mass per unit area of pool
}

