% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qh.R
\name{qh}
\alias{qh}
\title{Horizontal Darcy Velocity}
\usage{
qh(x, y = NULL, L, mfdata)
}
\arguments{
\item{x, y}{location, read by \code{xy.coords};
should be in absolute co-ordinates with the same origin and units as
 \code{mfdata}}

\item{L}{integer \code{[]};
a continuous range is insisted upon and this is corrected to
 \code{min(L):max(L)} within function}

\item{mfdata}{NetCDF object;
a MODFLOW dataset; see \code{\link[Rflow]{GW.nc}}}
}
\value{
numeric array \code{[length(L), NTS]}, dropped, where \code{NTS} is the
 number of time steps in the MODFLOW results;
the Darcy velocity in each layer for each time step at the location x,y
}
\description{
Calculates the horizontal Darcy velocity, or specific discharge, at a
 specific location in a MODFLOW model.
}
\details{
Uses linear interpolation to calculate the velocity between cell faces.
 For the layer with the water table, the thickness (and therefore flow
 normal area) is transient.  NA is returned for dry layers.
}
\examples{
library("RNetCDF")

# get example MODFLOW data set
mfdata <- open.nc(system.file("rflow_mf_demo.nc", package = "Rflow"))
qh(625, 825, 1L, mfdata)

}

