% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UK.to.site}
\alias{UK.to.site}
\title{Convert national usage trend to site usage rate through time}
\usage{
UK.to.site(nat.uhist, pu)
}
\arguments{
\item{nat.uhist}{data.frame or one of \code{"TCEusage"}, \code{"PCEusage"},
 \code{"TCAusage"} or \code{"TeCMusage"};
data frame must have one column called \code{cons}, which will be read
 as the usage rate}

\item{pu}{numeric \code{[1]};
the peak usage rate at the site ever - the national usage rate is
 normalised to give it a maximum value of \code{pu}; because \code{cons}
 is normalised, \code{pu} needn't be in the same units - it should be in
 the desired units of output and kg/day is suggested}
}
\value{
data.frame with adapted \code{cons} column
}
\description{
Convert national usage trend to site usage rate through time
}
\examples{
UK.to.site("TCEusage", 10)

}

