% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usage}
\alias{PCEusage}
\alias{TCAusage}
\alias{TCEusage}
\alias{TeCMusage}
\alias{usage}
\title{UK Chlorinated solvent usage}
\format{Data frames with 15 rows:
\describe{
  \item{year}{}
  \item{cons}{estimated total UK consumption, in kt/a (1000 metric
   tonnes per year)}
}}
\source{
\url{http://linkinghub.elsevier.com/retrieve/pii/002216949090181V}
}
\usage{
TCEusage

PCEusage

TCAusage

TeCMusage
}
\description{
Estimated chlorinated solvent UK-wide consumption over time in the
 twentieth century.
}
\details{
TCE: trichloroethene\cr
PCE: tetrachloroethene\cr
TCA: 1,1,1-trichloroethane\cr
TeCM: tetrachloromethane/ carbon tetrachloride

Check \code{attr(TCEusage, "units")} to see units displayed.
}
\references{
Rivett, M.O. et al., 1990. Organic contamination of the Birmingham aquifer, U.K. Journal of Hydrology, 113(1–4), pp.307–323.
}
\keyword{datasets}

