% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hL.R
\name{hL.setup}
\alias{hL.setup}
\title{Make DNAPL model layer heights appropriate to a MODFLOW data set}
\usage{
hL.setup(mfdata, x, y = NULL, ndlpmfl = 1L, bph = 0, z0 = "base",
  bph.warning = TRUE)
}
\arguments{
\item{mfdata}{NetCDF; a MODFLOW data set (see \code{\link[Rflow]{GW.nc}})}

\item{x, y}{for \code{\link[grDevices]{xy.coords}}; location of the source zone, in
the same co-ordinate system as \code{mfdata}; one location only}

\item{ndlpmfl}{integer \code{[1]} or \code{[mfNLAY]}, where \code{mfNLAY} is the number
 of layers in the MODFLOW model;
number of DNAPL model layers per MODFLOW layer (including any which are
 always dry at the location, although the values for these layers won't
 make any difference)}

\item{z0}{numeric \code{[1]} or \code{"base"};
elevation of the bottom of the DNAPL source zone, with the same datum as
 \code{mfdata}; \code{"base"} instructs to read the bottom of the
 MODFLOW model at this location}
}
\value{
numeric;
heights of layers of the DNAPL model, from top to bottom
}
\description{
Make DNAPL model layer heights appropriate to a MODFLOW data set
}
\examples{
mfdata <- RNetCDF::open.nc(system.file("rflow_mf_demo.nc",
                                       package = "Rflow"))

hL.setup(mfdata, 625, 825, 1L)
hL.setup(mfdata, 625, 825, 3L)
hL.setup(mfdata, 625, 825, 3L, z0 = 10)

# include a height-1 basal pool
hL.setup(mfdata, 625, 825, 3L, 1)

}
