% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UK.to.site}
\alias{UK.to.site}
\title{Convert national usage trend to site usage rate through time}
\usage{
UK.to.site(nat.uhist, pu, out.units = "kg/day")
}
\arguments{
\item{nat.uhist}{data.frame or one of \code{"TCEusage"}, \code{"PCEusage"},
 \code{"TCAusage"} or \code{"TeCMusage"};
data frame must have one column called \code{cons}, which will be read
 as the usage rate}

\item{pu}{numeric \code{[1]};
the peak usage rate at the site ever - the national usage rate is
 normalised to give it a maximum value of \code{pu}}

\item{out.units}{character string;
the units for the output data frame; cons values are
 multiplied appropriately depending on the units; permitted values are
 \code{"kt/year"}, \code{"t/year"}, \code{"kg/year"}, \code{"t/day"} and
 \code{"kg/day"}; the default is set up to be correct for the usage
 data sets (\code{TCEusage} etc.) and the DNST solver defaults.}
}
\value{
data.frame with adapted \code{cons} column
}
\description{
Convert national usage trend to site usage rate through time
}
\examples{
UK.to.site("TCEusage", 10)

}

